Only accept NaiveDate in penultimate calculation!  v0.1.2
We previously accepted a DateTime<Tz> generic timezone-aware date-time
object for the calculation of the penultimate day, but we immediately
discard the time information.  This resulted in messy code in which we
had to convert the time-aware datetime structure into a naive date
object multiple times, as well as effectively unnecessary imports.
This version cleans up this implementation a bit by only handling the
NaiveDate object.